backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
lightgbm==4.1.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2

